Fix ensure-collection name for Windows.
I would've preferred to change the colons to something else everywhere, but alas that would break existing systems on Linux. So the choice is either make a Windows exception, implement upgrading semantics, or allow bot hat the same time. The latter two would incur a lot more lambdalite hacks, which I'd rather not commit to.