Fixed problems with server interfaces (#14)
* Fixed starting wookie server

* i-wookie: Fixed request uri parsing and request header handling

* i-wookie: Properly add mime-type to static files

* i-wookie: Properly add mime-type to response from page handler

* i-woo: Corrected cookie parsing

so that it will accept extra '=' characters within value.

* i-woo: fixed outputting cookie header to response

* i-hunchentoot: make sure to read unprocessed body before sending response

In the case that post method has a body but for some reason it was
not read then read it before sending response to prevent the unread
body from messing the next request if connection keep-alive is in
effect.

* i-hunchentoot: optimize discarding unread body

and read only at most content length. If the request is not of the
two supported formats (form or multipart) then caller is expected
to handle body wholly or not at all.

* i-hunchentoot: no need to recreate discard buffer for each 4k loaded

* i-hunchentoot: streamlined reading the unprocessed body

based on suggestion by Shinmera. Two previous commits are also based
on his reviews.

* i-hunchentoot: fixed requests with no body

If there was no body then content length returned was nil not a number.

* i-woo & i-wookie: refactored a little bit

based on review comments.